update financial handling

+++ client/app/routes/erp+/_admin+/employees+/$employeeId.tsx
@@ -1,25 +1,33 @@
 import { deleteEmployee, getEmployeeById } from '~/services/employee.server';
 import { LoaderFunctionArgs, ActionFunctionArgs } from '@remix-run/node';
 import { isAuthenticated } from '~/services/auth.server';
-import { data, useLoaderData, useNavigate } from '@remix-run/react';
+import { data, redirect, useLoaderData, useNavigate } from '@remix-run/react';
 import { getLast7DaysStats } from '~/services/attendance.server';
-import { parseAuthCookie } from '~/services/cookie.server';
+import {
+  getUnauthorizedActionResponse,
+  parseAuthCookie,
+} from '~/services/cookie.server';
 import ContentHeader from '~/components/ContentHeader';
 import EmployeeDetail from './_components/EmployeeDetail';
 import EmployeeAttendanceList from './_components/EmployeeAttendanceList';
 import { Edit, Pen } from 'lucide-react';
 import { canAccessEmployeeManagement } from '~/utils/permission';
+import { IActionFunctionReturn } from '~/interfaces/app.interface';
+import { getTasks } from '~/services/task.server';
+import EmployeeTaskList from './_components/EmployeeTaskList';
+import { getFirstWeekDate } from '~/utils/date.util';
+import { TODAY } from '~/constants/date.constant';
 
 export const loader = async ({ request, params }: LoaderFunctionArgs) => {
   const session = await parseAuthCookie(request);
+  const employeeId = params.employeeId;
 
   if (!canAccessEmployeeManagement(session?.user.usr_role)) {
-    throw new Response('Bạn không có quyền truy cập vào trang này.', {
-      status: 403,
-    });
+    // throw new Response('Bạn không có quyền truy cập vào trang này.', {
+    //   status: 403,
+    // });
+    return redirect(`/erp/nhan-vien/employees/${employeeId}`);
   }
-
-  const employeeId = params.employeeId;
 
   if (!employeeId) {
     throw new Response('Employee ID is required', { status: 400 });
@@ -38,7 +46,13 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
   // Get employee first to get user ID for attendance
   const attendancePromise = getEmployeeById(employeeId, session!)
     .then((employee) => {
-      return getLast7DaysStats(employee.emp_user.id, session!);
+      return getLast7DaysStats(employee.emp_user.id, session!).catch((e) => {
+        console.error('Error fetching attendance stats:', e);
+        return {
+          success: false,
+          message: e.message || 'Có lỗi khi lấy thống kê chấm công',
+        };
+      });
     })
     .catch((error) => {
       console.error('Error fetching attendance stats:', error);
@@ -48,14 +62,33 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
       };
     });
 
+  const taskPromise = getTasks(
+    new URLSearchParams({
+      assignee: employeeId,
+      startDateFrom: getFirstWeekDate(
+        TODAY.getDay(),
+        TODAY.getMonth(),
+        TODAY.getFullYear(),
+      ).toISOString(),
+    }),
+    session!,
+  ).catch((error) => {
+    console.error('Error fetching tasks:', error);
+    return {
+      success: false,
+      message: error.message || 'Có lỗi khi lấy danh sách công việc',
+    };
+  });
+
   return {
     employeeId,
     employeePromise,
     attendancePromise,
+    taskPromise,
   };
 };
 export default function EmployeeDetails() {
-  const { employeeId, employeePromise, attendancePromise } =
+  const { employeeId, employeePromise, attendancePromise, taskPromise } =
     useLoaderData<typeof loader>();
   const navigate = useNavigate();
 
@@ -84,17 +117,26 @@ export default function EmployeeDetails() {
         employeeId={employeeId}
         attendancePromise={attendancePromise}
       />
+
+      {/* Employee task List Card */}
+      <EmployeeTaskList employeeId={employeeId} taskPromise={taskPromise} />
     </div>
   );
 }
 
-export const action = async ({ request, params }: ActionFunctionArgs) => {
+export const action = async ({
+  request,
+  params,
+}: ActionFunctionArgs): IActionFunctionReturn => {
+  const { session, headers } = await isAuthenticated(request);
+  if (!session) return getUnauthorizedActionResponse(data, headers);
+
   switch (request.method) {
     case 'DELETE':
-      const { session, headers } = await isAuthenticated(request);
       await deleteEmployee(params.employeeId!, session!);
       return data(
         {
+          success: true,
           toast: {
             type: 'success' as const,
             message: 'Xóa nhân viên thành công',
@@ -105,15 +147,13 @@ export const action = async ({ request, params }: ActionFunctionArgs) => {
     default:
       return data(
         {
+          success: false,
           toast: {
             type: 'error' as const,
             message: 'Phương thức không hợp lệ',
           },
         },
-        {
-          status: 405,
-          statusText: 'Method Not Allowed',
-        },
+        { headers, status: 405, statusText: 'Method Not Allowed' },
       );
   }
 };
